+++ src/components/layout/Header.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment happy-dom
-import Header from '$components/Header.svelte';
+import Header from '$components/layout/Header.svelte';
 import { render } from '@testing-library/svelte';
 import { expect, it } from 'vitest';
