+++ src/routes/+layout.svelte
@@ -2,8 +2,8 @@
 	// TODO: Dynamic theme selection; re-render layout based on key
 	import 'highlight.js/styles/atom-one-dark-reasonable.css';
 
-	import Footer from '$components/Footer.svelte';
-	import Header from '$components/Header.svelte';
+	import Footer from '$components/layout/Footer.svelte';
+	import Header from '$components/layout/Header.svelte';
 	import { fade } from 'svelte/transition';
 	import '../app.css';
